move bool types to real modules
