+++ crates/libs/windows/src/Windows/Foundation/mod.rs
@@ -3216,6 +3216,5 @@ where
     pub TSender: core::marker::PhantomData<TSender>,
     pub TResult: core::marker::PhantomData<TResult>,
 }
-core::include!(core::concat!(core::env!("CARGO_MANIFEST_DIR"), "/src/includes/", "Foundation/TimeSpan.rs"));
 #[cfg(feature = "implement")]
 core::include!("impl.rs");
