+++ crates/libs/windows/src/extensions.rs
@@ -1,0 +1,9 @@
+//! Contains extensions to the generated bindings
+//!
+//! This module contains code that extends the functionality of the bindings generated by `bindgen`.
+//! It adds inherent impls, trait impls, etc.
+
+#[cfg(feature = "Foundation")]
+pub mod Foundation;
+#[cfg(feature = "Win32")]
+pub mod Win32;
