+++ crates/libs/windows/src/extensions/Foundation.rs
@@ -1,0 +1,1 @@
+pub mod TimeSpan;
