+++ crates/libs/windows/src/extensions/Win32.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "Win32_Foundation")]
+pub mod Foundation;
