+++ crates/libs/windows/src/extensions/Win32/Foundation.rs
@@ -1,0 +1,5 @@
+pub mod BOOL;
+pub mod BOOLEAN;
+pub mod NTSTATUS;
+pub mod VARIANT_BOOL;
+pub mod WIN32_ERROR;
