+++ crates/libs/windows/src/extensions/Win32/Foundation/NTSTATUS.rs
@@ -1,3 +1,5 @@
+use crate::Win32::Foundation::NTSTATUS;
+
 impl NTSTATUS {
     #[inline]
     pub const fn is_ok(self) -> bool {
